requirements.txt
Keras==2.0.5
tensorflow==1.2.0
Flask==0.12.2
matplotlib==1.3.1
Pillow==4.1.1
opencv-python==3.2.0.7
tqdm==4.14.0